Reprocess endpoint (#387)
* reprocess wip

* fix tests

* remove old data before writing new

* update description

+++ stixify/worker/tasks.py
@@ -2,6 +2,7 @@ from datetime import UTC, datetime
 import logging
 import os
 from pathlib import Path
+import uuid
 from stixify.web.models import Job, File
 from stixify.web import models
 from celery import shared_task
@@ -11,10 +12,12 @@ from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.core.files.storage import default_storage
 from django.core.files.base import File as DjangoFile
 from django.core.files.base import File as DjangoFile
+from django.db import transaction
 import stix2
 
 from stixify.worker import helpers, pdf_converter
 from django.conf import settings
+from txt2stix.txt2stix import Txt2StixData
 
 
 POLL_INTERVAL = 1
@@ -25,6 +28,17 @@ def new_task(job: Job):
         countdown=POLL_INTERVAL, root_id=str(job.id), task_id=str(job.id)
     )
 
+def create_reprocessing_job(file: File, options: dict = None):
+    options = options or {}
+    job  = models.Job.objects.create(
+        id=uuid.uuid4(),
+        type=models.JobType.REPROCESS_POSTS,
+        file=file,
+        state=models.JobState.PENDING,
+        extra=options,
+    )
+    new_task(job)
+    return job
 
 @shared_task
 def process_post(job_id, *args):
@@ -66,32 +80,45 @@ def process_post(job_id, *args):
         processor.setup(
             report_prop=report_props, extra=dict(_stixify_file_id=str(file.id))
         )
-        processor.process()
-        if processor.incident:
-            file.ai_describes_incident = processor.incident.describes_incident
-            file.ai_incident_summary = processor.incident.explanation
-            file.ai_incident_classification = processor.incident.incident_classification
+        skip_extraction = bool((job.extra or {}).get("skip_extraction"))
 
-        file.txt2stix_data = processor.txt2stix_data.model_dump(
-            mode="json", exclude_defaults=True, exclude_unset=True, exclude_none=True
-        )
-        file.summary = processor.summary
-        file.markdown_file.save("markdown.md", processor.md_file.open(), save=True)
+        # remove existing values for this file that are not in the new upload (handles deletions and modifications)
+        models.ObjectValue.objects.filter(file_id=file.id).delete()
+        if job.type == models.JobType.REPROCESS_POSTS and skip_extraction:
+            processor.output_md = file.markdown_file.open().read().decode()
+            txt2stix_data = None
+            if not file.txt2stix_data:
+                raise Exception("no existing extraction data to use for reprocess with skip_extraction=true")
+            txt2stix_data = Txt2StixData.model_validate(file.txt2stix_data)
+            processor.txt2stix(txt2stix_data)
+            processor.write_bundle(processor.bundler)
+            processor.upload_to_arango()
+        else:
+            processor.process()
+        
+        with transaction.atomic(): # revert to old file if something goes wrong during processing
+            new_profile_id = (job.extra or {}).get("profile_id")
+            if new_profile_id:
+                file.profile_id = new_profile_id
+                file.save(update_fields=["profile"])
+            file.set_txt2stix_data(processor.txt2stix_data)
+            file.create_embedding(include_non_incident=settings.CREATE_EMBEDDING_INCLUDE_NON_INCIDENT)
 
-        models.FileImage.objects.filter(report=file).delete()  # remove old references
+            if job.type == models.JobType.IMPORT_FILE: # only update files for import jobs, reprocess jobs should keep the same file references
+                file.markdown_file.save("markdown.md", processor.md_file.open(), save=True)
+                models.FileImage.objects.filter(report=file).delete()  # remove old references
 
-        for image in processor.md_images:
-            models.FileImage.objects.create(
-                report=file, file=DjangoFile(image, image.name), name=image.name
-            )
+                for image in processor.md_images:
+                    models.FileImage.objects.create(
+                        report=file, file=DjangoFile(image, image.name), name=image.name
+                    )
 
-        converted_file_path = processor.tmpdir / "converted_pdf.pdf"
-        pdf_converter.make_conversion(processor.filename, converted_file_path)
-        file.create_embedding(include_non_incident=settings.CREATE_EMBEDDING_INCLUDE_NON_INCIDENT)
-        file.pdf_file.save(
-            converted_file_path.name, open(converted_file_path, mode="rb")
-        )
-        file.save()
+                converted_file_path = processor.tmpdir / "converted_pdf.pdf"
+                pdf_converter.make_conversion(processor.filename, converted_file_path)
+                file.pdf_file.save(
+                    converted_file_path.name, open(converted_file_path, mode="rb")
+                )
+                file.save(update_fields=['markdown_file', 'pdf_file'])
     except Exception as e:
         error = str(e)
         job.error = "failed to process report"
@@ -109,7 +136,8 @@ def job_completed_with_error(job_id):
     state = models.JobState.COMPLETED
     if job.error:
         state = models.JobState.FAILED
-        job.file and job.file.delete()
+        if job.type == models.JobType.IMPORT_FILE:
+            job.file and job.file.delete()
     Job.objects.filter(pk=job_id).update(state=state, completion_time=datetime.now(UTC))
 
 @shared_task
